LPS-56481 Move comment manager implementation and test into modules
