+++ modules/apps/message-boards/message-boards-comment-impl/src/com/liferay/message/boards/comment/impl/MBDiscussionPermissionImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.messageboards.comment;
+package com.liferay.message.boards.comment.impl;
 
 import com.liferay.portal.kernel.comment.BaseDiscussionPermission;
 import com.liferay.portal.kernel.exception.PortalException;
