+++ modules/apps/message-boards/message-boards-comment-impl/src/com/liferay/message/boards/comment/impl/MBDiscussionStagingHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.messageboards.comment;
+package com.liferay.message.boards.comment.impl;
 
 import com.liferay.portal.kernel.comment.DiscussionStagingHandler;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
